pyproject.toml
[tool.poetry]
name = "y42-modeling-test"
version = "0.1.0"
description = ""
authors = ["Adrien Carpentier <moi@adriencarpentier.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tortoise-orm = "^0.17.7"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
mypy = "^0.941"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
